Added more advanced control to the Arm subsystem
I added several methods to the Arm subsystem to more fully control the arm.

set was renamed to setExtension, as it controls the extending motors

setShoulder controls the lifting of the arm from the shoulder joint.

toPosition is the highest level of control.  Given an x and y coord in the from of a Translation2d, it moves the arm to the desired position

toExtension extends the arm to the desired distance in meters

toAngle lifts the arm to the desired angle above 0, which is when the arm is facing straight down.

There are also 3 getter methods for that do the inverse of the three "to_____" methods.